requirements.txt
Flask
ultralytics
torch
torchvision
Pillow
opencv-python
numpy